Added Spring WebFilter to forward certain routes to react-router
Spring will now forward requests matching react-router routes to the appropriate
client-side route.

Q: Why do we need to do this?
A: Because Spring is not aware of the routes that we configure in react-router;
these are client-side routes, not actual routes that exist on the server. Thus,
we need Spring to forward requests made to these client-side routes back to the
client on `index.html` where `react-router` can work its magic.

Related work items: #41